Add nn.Function
