+++ configs/csg/pbe_with_repl_pretrain_base.yaml
@@ -69,8 +69,10 @@ loss_fn:
           constants:
             rhs: "@/batch_size"
       - - "pick"
-        - type: mlprogram.nn.Pick
-          key: action_sequence_loss
+        - type: mlprogram.nn.Function
+          f:
+            type: mlprogram.utils.Pick
+            key: action_sequence_loss
 main:
   type: mlprogram.entrypoint.train_supervised
   workspace_dir: "output/workspace_pretraining"
