+++ mlprogram/nn/__init__.py
@@ -6,5 +6,5 @@ from mlprogram.nn.cnn import CNN2d  # noqa
 from mlprogram.nn.mlp import MLP  # noqa
 from mlprogram.nn.apply import Apply  # noqa
 from mlprogram.nn.aggregated_loss import AggregatedLoss # noqa
-from mlprogram.nn.pick import Pick  # noqa
+from mlprogram.nn.function import Function  # noqa
 from mlprogram.nn.primitives import Add, Sub, Mul, Div, IntDiv, Neg  # noqa
