+++ test/test_nl2code.py
@@ -126,7 +126,9 @@ class TestNL2Code(unittest.TestCase):
             to_action_sequence = AstToActionSequence()
             loss_fn = nn.Sequential(OrderedDict([
                 ("loss", Loss()),
-                ("pick", mlprogram.nn.Pick("action_sequence_loss"))
+                ("pick",
+                 mlprogram.nn.Function(
+                     mlprogram.utils.Pick("action_sequence_loss")))
             ]))
             collate = Collate(
                 torch.device("cpu"),
